+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 Changelog
 =========
 
+### 0.6.2
+
+* Should fix ide freezes with class autocompletion (use XXX).
+
+### 0.6.1
+
+* Service aliases support
+* Resolve services classes case insensitively
+
 ### 0.6.0
 
 * Autocomplete route name in php and twig